bin edge *.9 hack to avoid double counting
If I use integration limits that match the bin edges (e.g. 20.0,
30.0), the values of first(last)_pt_bin provided by TAxis::FindFixBin
are +1 too high. This leads to some pt1 slices to be counted twice. I
don't know whether this is a 'feature' of `FindFixBin` (see
e.g. https://root.cern.ch/phpBB3/viewtopic.php?t=8806) or just an
issue with float comparison mixed with python <--> c++ translation.
Anyway, this hack is good enough to get the values I need (which now
add up to the correct total).
Bugfix on 3768080